refactor SmartAction and MoveRunner
