+++ smart-actions/src/main/java/org/smartdata/actions/hdfs/BlockErasureCodeFileAction.java
@@ -19,6 +19,7 @@ package org.smartdata.actions.hdfs;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.smartdata.actions.ActionException;
 
 /**
  * An action to do block level erasure code a file, for both Hadoop 2.x and Hadoop 3.x.
@@ -27,6 +28,6 @@ public class BlockErasureCodeFileAction extends HdfsAction {
   private static final Logger LOG = LoggerFactory.getLogger(BlockErasureCodeFileAction.class);
 
   @Override
-  protected void execute() {
+  protected void execute() throws ActionException {
   }
 }
