+++ smart-actions/src/main/java/org/smartdata/actions/hdfs/StripErasureCodeFileAction.java
@@ -19,6 +19,7 @@ package org.smartdata.actions.hdfs;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.smartdata.actions.ActionException;
 
 /**
  * An action to do strip level erasure code a file. Only for Hadoop 3.x.
@@ -27,6 +28,6 @@ public class StripErasureCodeFileAction extends HdfsAction {
   private static final Logger LOG = LoggerFactory.getLogger(StripErasureCodeFileAction.class);
 
   @Override
-  protected void execute() {
+  protected void execute() throws ActionException {
   }
 }
